+++ lib/app.dart
@@ -14,3 +14,4 @@ library app;
 export 'src/app/moose_app_context.dart';
 export 'src/app/moose_scope.dart';
 export 'src/app/moose_bootstrapper.dart';
+export 'src/app/moose_lifecycle_observer.dart';